+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/SwapConfirmViewModel.kt
@@ -7,6 +7,8 @@ import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.ViewModelUiState
 import io.horizontalsystems.bankwallet.core.managers.CurrencyManager
 import io.horizontalsystems.bankwallet.entities.Currency
+import io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction.ISendTransactionService
+import io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction.SendTransactionServiceFactory
 import io.horizontalsystems.marketkit.models.Token
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
