+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/providers/BaseUniswapV3Provider.kt
@@ -2,8 +2,8 @@ package io.horizontalsystems.bankwallet.modules.swapxxx.providers
 
 import io.horizontalsystems.bankwallet.modules.swapxxx.EvmBlockchainHelper
 import io.horizontalsystems.bankwallet.modules.swapxxx.ISwapQuote
-import io.horizontalsystems.bankwallet.modules.swapxxx.SendTransactionData
 import io.horizontalsystems.bankwallet.modules.swapxxx.SwapQuoteUniswapV3
+import io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction.SendTransactionData
 import io.horizontalsystems.bankwallet.modules.swapxxx.settings.SwapSettingDeadline
 import io.horizontalsystems.bankwallet.modules.swapxxx.settings.SwapSettingRecipient
 import io.horizontalsystems.bankwallet.modules.swapxxx.settings.SwapSettingSlippage
