+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/providers/ISwapXxxProvider.kt
@@ -1,7 +1,7 @@
 package io.horizontalsystems.bankwallet.modules.swapxxx.providers
 
 import io.horizontalsystems.bankwallet.modules.swapxxx.ISwapQuote
-import io.horizontalsystems.bankwallet.modules.swapxxx.SendTransactionData
+import io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction.SendTransactionData
 import io.horizontalsystems.marketkit.models.BlockchainType
 import io.horizontalsystems.marketkit.models.Token
 import java.math.BigDecimal
