+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/sendtransaction/SendTransactionData.kt
@@ -1,0 +1,7 @@
+package io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction
+
+import io.horizontalsystems.ethereumkit.models.TransactionData
+
+sealed class SendTransactionData {
+    data class Evm(val transactionData: TransactionData): SendTransactionData()
+}
