+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/sendtransaction/SendTransactionServiceEvm.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.swapxxx
+package io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction
 
 import androidx.compose.runtime.Composable
 import androidx.lifecycle.ViewModel
@@ -22,41 +22,6 @@ import io.horizontalsystems.ethereumkit.core.LegacyGasPriceProvider
 import io.horizontalsystems.ethereumkit.core.eip1559.Eip1559GasPriceProvider
 import io.horizontalsystems.ethereumkit.models.TransactionData
 import io.horizontalsystems.marketkit.models.BlockchainType
-import io.horizontalsystems.marketkit.models.Token
-
-object SendTransactionServiceFactory {
-    fun create(tokenIn: Token): ISendTransactionService? = when (val blockchainType = tokenIn.blockchainType) {
-        BlockchainType.Bitcoin -> TODO()
-        BlockchainType.BitcoinCash -> TODO()
-        BlockchainType.ECash -> TODO()
-        BlockchainType.Litecoin -> TODO()
-        BlockchainType.Dash -> TODO()
-        BlockchainType.Zcash -> TODO()
-        BlockchainType.Ethereum -> SendTransactionServiceEvm(blockchainType)
-        BlockchainType.BinanceSmartChain -> SendTransactionServiceEvm(blockchainType)
-        BlockchainType.BinanceChain -> TODO()
-        BlockchainType.Polygon -> TODO()
-        BlockchainType.Avalanche -> TODO()
-        BlockchainType.Optimism -> TODO()
-        BlockchainType.ArbitrumOne -> TODO()
-        BlockchainType.Solana -> TODO()
-        BlockchainType.Gnosis -> TODO()
-        BlockchainType.Fantom -> TODO()
-        BlockchainType.Tron -> TODO()
-        BlockchainType.Ton -> TODO()
-        is BlockchainType.Unsupported -> null
-    }
-}
-
-interface ISendTransactionService {
-    @Composable
-    fun GetContent(navController: NavController)
-    fun setSendTransactionData(data: SendTransactionData)
-}
-
-sealed class SendTransactionData {
-    data class Evm(val transactionData: TransactionData): SendTransactionData()
-}
 
 class SendTransactionServiceEvm(blockchainType: BlockchainType) : ISendTransactionService {
     private lateinit var transactionData: TransactionData
