+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/sendtransaction/SendTransactionServiceFactory.kt
@@ -1,0 +1,28 @@
+package io.horizontalsystems.bankwallet.modules.swapxxx.sendtransaction
+
+import io.horizontalsystems.marketkit.models.BlockchainType
+import io.horizontalsystems.marketkit.models.Token
+
+object SendTransactionServiceFactory {
+    fun create(tokenIn: Token): ISendTransactionService? = when (val blockchainType = tokenIn.blockchainType) {
+        BlockchainType.Bitcoin -> TODO()
+        BlockchainType.BitcoinCash -> TODO()
+        BlockchainType.ECash -> TODO()
+        BlockchainType.Litecoin -> TODO()
+        BlockchainType.Dash -> TODO()
+        BlockchainType.Zcash -> TODO()
+        BlockchainType.Ethereum -> SendTransactionServiceEvm(blockchainType)
+        BlockchainType.BinanceSmartChain -> SendTransactionServiceEvm(blockchainType)
+        BlockchainType.BinanceChain -> TODO()
+        BlockchainType.Polygon -> TODO()
+        BlockchainType.Avalanche -> TODO()
+        BlockchainType.Optimism -> TODO()
+        BlockchainType.ArbitrumOne -> TODO()
+        BlockchainType.Solana -> TODO()
+        BlockchainType.Gnosis -> TODO()
+        BlockchainType.Fantom -> TODO()
+        BlockchainType.Tron -> TODO()
+        BlockchainType.Ton -> TODO()
+        is BlockchainType.Unsupported -> null
+    }
+}
